move mol to formatters dir
